Refactored to maintain modularity.
