+++ src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './components/auth/login.component';
-import { SignupComponent } from './components/auth/signup.component';
-import { LandingComponent } from './components/landing.component';
+import { LoginComponent } from './features/auth/components/login/login.component';
+import { SignupComponent } from './features/auth/components/signup/signup.component';
+import { LandingComponent } from './features/landing-page/components/landing.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
