--- src/app/components/auth/login.component.ts
@@ -1,258 +1,0 @@
-import { Component, signal } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
-import { Router, RouterModule } from '@angular/router';
-import { AuthService } from '../../services/auth.service';
-import { FormInputComponent } from '../ui/form-input.component';
-import { ButtonComponent } from '../ui/button.component';
-
-@Component({
-  selector: 'app-login',
-  standalone: true,
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    RouterModule,
-    FormInputComponent,
-    ButtonComponent
-  ],
-  template: `
-    <div class="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-4">
-      <div class="w-full max-w-md">
-        <!-- Header -->
-        <div class="text-center mb-8">
-          <div class="w-16 h-16 bg-gradient-to-br from-blue-500 to-blue-700 rounded-2xl mx-auto mb-4 flex items-center justify-center shadow-lg">
-            <svg class="w-8 h-8 text-white" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 16l-4-4m0 0l4-4m-4 4h14m-5 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h7a3 3 0 013 3v1"/>
-            </svg>
-          </div>
-          <h1 class="text-3xl font-bold text-gray-900 mb-2">Welcome Back</h1>
-          <p class="text-gray-600">Sign in to your UnravelDocs account</p>
-        </div>
-
-        <!-- Form -->
-        <div class="bg-white rounded-2xl shadow-xl p-8">
-          <form [formGroup]="loginForm" (ngSubmit)="onSubmit()" class="space-y-6">
-            <!-- Email -->
-            <app-form-input
-              id="email"
-              label="Email Address"
-              type="email"
-              placeholder="Enter your email address"
-              formControlName="email"
-              [required]="true"
-              [hasError]="isFieldInvalid('email')"
-              [errorMessage]="getFieldError('email')"
-              icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207"/></svg>'
-              autocomplete="email"
-            />
-
-            <!-- Password -->
-            <app-form-input
-              id="password"
-              label="Password"
-              type="password"
-              placeholder="Enter your password"
-              formControlName="password"
-              [required]="true"
-              [hasError]="isFieldInvalid('password')"
-              [errorMessage]="getFieldError('password')"
-              icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"/></svg>'
-              autocomplete="current-password"
-            />
-
-            <!-- Remember Me and Forgot Password -->
-            <div class="flex items-center justify-between">
-              <div class="flex items-center gap-3">
-                <div class="relative">
-                  <input
-                    id="rememberMe"
-                    type="checkbox"
-                    formControlName="rememberMe"
-                    class="sr-only"
-                  >
-                  <div
-                    class="w-5 h-5 border-2 rounded transition-all duration-200 flex items-center justify-center cursor-pointer"
-                    [class]="loginForm.get('rememberMe')?.value
-                      ? 'bg-blue-600 border-blue-600'
-                      : 'bg-white border-gray-300 hover:border-blue-400'"
-                    (click)="toggleRememberMe()"
-                  >
-                    <svg *ngIf="loginForm.get('rememberMe')?.value" class="w-3 h-3 text-white" fill="currentColor" viewBox="0 0 20 20">
-                      <path fill-rule="evenodd" d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z" clip-rule="evenodd"/>
-                    </svg>
-                  </div>
-                </div>
-                <label for="rememberMe" class="text-sm text-gray-700 cursor-pointer">
-                  Remember me
-                </label>
-              </div>
-
-              <a
-                routerLink="/forgot-password"
-                class="text-sm text-blue-600 hover:text-blue-800 font-medium transition-colors"
-              >
-                Forgot password?
-              </a>
-            </div>
-
-            <!-- Error Message -->
-            <div *ngIf="errorMessage()" class="p-4 bg-red-50 border border-red-200 rounded-xl">
-              <div class="flex items-center gap-3">
-                <svg class="w-5 h-5 text-red-500 flex-shrink-0" fill="currentColor" viewBox="0 0 20 20">
-                  <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"/>
-                </svg>
-                <p class="text-sm text-red-800">{{ errorMessage() }}</p>
-              </div>
-            </div>
-
-            <!-- Submit Button -->
-            <app-button
-              type="submit"
-              variant="primary"
-              size="lg"
-              class="w-full"
-              [disabled]="loginForm.invalid || isLoading()"
-              [loading]="isLoading()"
-            >
-              <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M11 16l-4-4m0 0l4-4m-4 4h14m-5 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013 3v1"/>
-              </svg>
-              Sign In
-            </app-button>
-          </form>
-
-          <!-- Signup Link -->
-          <div class="mt-8 text-center">
-            <p class="text-gray-600">
-              Don't have an account?
-              <a routerLink="/signup" class="text-blue-600 hover:text-blue-800 font-semibold transition-colors">
-                Create one here
-              </a>
-            </p>
-          </div>
-        </div>
-
-        <!-- Social Login -->
-        <div class="mt-8">
-          <div class="relative">
-            <div class="absolute inset-0 flex items-center">
-              <div class="w-full border-t border-gray-300"></div>
-            </div>
-            <div class="relative flex justify-center text-sm">
-              <span class="px-4 bg-gray-50 text-gray-500">Or continue with</span>
-            </div>
-          </div>
-
-          <div class="mt-6 grid grid-cols-2 gap-3">
-            <button
-              type="button"
-              class="flex justify-center items-center gap-2 w-full px-4 py-3 border border-gray-300 rounded-xl shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-              (click)="loginWithGoogle()"
-            >
-              <svg class="w-5 h-5" viewBox="0 0 24 24">
-                <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-                <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-                <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-                <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-              </svg>
-              Google
-            </button>
-
-            <button
-              type="button"
-              class="flex justify-center items-center gap-2 w-full px-4 py-3 border border-gray-300 rounded-xl shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-              (click)="loginWithGitHub()"
-            >
-              <svg class="w-5 h-5" fill="currentColor" viewBox="0 0 24 24">
-                <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
-              </svg>
-              GitHub
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  `
-})
-export class LoginComponent {
-  loginForm: FormGroup;
-  errorMessage = signal('');
-  isLoading = signal(false);
-
-  constructor(
-    private fb: FormBuilder,
-    private authService: AuthService,
-    private router: Router
-  ) {
-    this.loginForm = this.fb.group({
-      email: ['', [Validators.required, Validators.email]],
-      password: ['', [Validators.required, Validators.minLength(6)]],
-      rememberMe: [false]
-    });
-  }
-
-  isFieldInvalid(fieldName: string): boolean {
-    const field = this.loginForm.get(fieldName);
-    return !!(field?.invalid && (field?.dirty || field?.touched));
-  }
-
-  getFieldError(fieldName: string): string {
-    const field = this.loginForm.get(fieldName);
-    if (field?.errors && (field.dirty || field.touched)) {
-      if (field.errors['required']) return `${this.getFieldLabel(fieldName)} is required`;
-      if (field.errors['email']) return 'Please enter a valid email address';
-      if (field.errors['minlength']) return `${this.getFieldLabel(fieldName)} must be at least ${field.errors['minlength'].requiredLength} characters`;
-    }
-    return '';
-  }
-
-  private getFieldLabel(fieldName: string): string {
-    const labels: { [key: string]: string } = {
-      email: 'Email address',
-      password: 'Password'
-    };
-    return labels[fieldName] || fieldName;
-  }
-
-  toggleRememberMe(): void {
-    const current = this.loginForm.get('rememberMe')?.value;
-    this.loginForm.patchValue({ rememberMe: !current });
-  }
-
-  async onSubmit(): Promise<void> {
-    if (this.loginForm.invalid) {
-      this.loginForm.markAllAsTouched();
-      return;
-    }
-
-    this.isLoading.set(true);
-    this.errorMessage.set('');
-
-    try {
-      const formValue = this.loginForm.value;
-      await this.authService.login({
-        email: formValue.email,
-        password: formValue.password,
-        rememberMe: formValue.rememberMe
-      });
-
-      this.router.navigate(['/dashboard']);
-    } catch (error: any) {
-      this.errorMessage.set(error.message || 'Invalid email or password. Please try again.');
-    } finally {
-      this.isLoading.set(false);
-    }
-  }
-
-  async loginWithGoogle(): Promise<void> {
-    // Implement Google OAuth login
-    console.log('Login with Google');
-  }
-
-  async loginWithGitHub(): Promise<void> {
-    // Implement GitHub OAuth login
-    console.log('Login with GitHub');
-  }
-}
-
