--- src/app/components/auth/signup.component.ts
@@ -1,384 +1,0 @@
-import { Component, signal } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
-import { Router, RouterModule } from '@angular/router';
-import { AuthService } from '../../services/auth.service';
-import { FormInputComponent } from '../ui/form-input.component';
-import { ButtonComponent } from '../ui/button.component';
-import { PasswordStrengthComponent } from '../ui/password-strength.component';
-import { PasswordGeneratorComponent } from '../ui/password-generator.component';
-
-@Component({
-  selector: 'app-signup',
-  standalone: true,
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    RouterModule,
-    FormInputComponent,
-    ButtonComponent,
-    PasswordStrengthComponent,
-    PasswordGeneratorComponent
-  ],
-  template: `
-    <div class="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-4">
-      <div class="w-full max-w-md">
-        <!-- Header -->
-        <div class="text-center mb-8">
-          <div class="w-16 h-16 bg-gradient-to-br from-blue-500 to-blue-700 rounded-2xl mx-auto mb-4 flex items-center justify-center shadow-lg">
-            <svg class="w-8 h-8 text-white" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M18 9v3m0 0v3m0-3h3m-3 0h-3m-2-5a4 4 0 11-8 0 4 4 0 018 0zM3 20a6 6 0 0112 0v1H3v-1z"/>
-            </svg>
-          </div>
-          <h1 class="text-3xl font-bold text-gray-900 mb-2">Create Account</h1>
-          <p class="text-gray-600">Join UnravelDocs and start organizing your documents</p>
-        </div>
-
-        <!-- Form -->
-        <div class="bg-white rounded-2xl shadow-xl p-8">
-          <form [formGroup]="signupForm" (ngSubmit)="onSubmit()" class="space-y-6">
-            <!-- Full Name -->
-            <app-form-input
-              id="fullName"
-              label="Full Name"
-              type="text"
-              placeholder="Enter your full name"
-              formControlName="fullName"
-              [required]="true"
-              [hasError]="isFieldInvalid('fullName')"
-              [errorMessage]="getFieldError('fullName')"
-              icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"/></svg>'
-              autocomplete="name"
-            />
-
-            <!-- Email -->
-            <app-form-input
-              id="email"
-              label="Email Address"
-              type="email"
-              placeholder="Enter your email address"
-              formControlName="email"
-              [required]="true"
-              [hasError]="isFieldInvalid('email')"
-              [errorMessage]="getFieldError('email')"
-              icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207"/></svg>'
-              autocomplete="email"
-            />
-
-            <!-- Password -->
-            <div>
-              <app-form-input
-                id="password"
-                label="Password"
-                type="password"
-                placeholder="Create a secure password"
-                formControlName="password"
-                [required]="true"
-                [hasError]="isFieldInvalid('password')"
-                [errorMessage]="getFieldError('password')"
-                [showGenerateButton]="true"
-                (generatePassword)="showPasswordGenerator()"
-                icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"/></svg>'
-                autocomplete="new-password"
-                (input)="onPasswordChange($event)"
-              />
-
-              <!-- Password Strength Indicator -->
-              <app-password-strength [password]="currentPassword()"/>
-            </div>
-
-            <!-- Confirm Password -->
-            <app-form-input
-              id="confirmPassword"
-              label="Confirm Password"
-              type="password"
-              placeholder="Confirm your password"
-              formControlName="confirmPassword"
-              [required]="true"
-              [hasError]="isFieldInvalid('confirmPassword')"
-              [errorMessage]="getFieldError('confirmPassword')"
-              icon='<svg fill="none" stroke="currentColor" viewBox="0 0 24 24"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"/></svg>'
-              autocomplete="new-password"
-            />
-
-            <!-- Terms and Conditions -->
-            <div class="flex items-start gap-3">
-              <div class="relative">
-                <input
-                  id="terms"
-                  type="checkbox"
-                  formControlName="terms"
-                  class="sr-only"
-                >
-                <div
-                  class="w-5 h-5 border-2 rounded transition-all duration-200 flex items-center justify-center cursor-pointer"
-                  [class]="signupForm.get('terms')?.value
-                    ? 'bg-blue-600 border-blue-600'
-                    : 'bg-white border-gray-300 hover:border-blue-400'"
-                  (click)="toggleTerms()"
-                >
-                  <svg *ngIf="signupForm.get('terms')?.value" class="w-3 h-3 text-white" fill="currentColor" viewBox="0 0 20 20">
-                    <path fill-rule="evenodd" d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z" clip-rule="evenodd"/>
-                  </svg>
-                </div>
-              </div>
-              <label for="terms" class="text-sm text-gray-700 cursor-pointer">
-                I agree to the
-                <a href="/terms" class="text-blue-600 hover:text-blue-800 font-medium">Terms of Service</a>
-                and
-                <a href="/privacy" class="text-blue-600 hover:text-blue-800 font-medium">Privacy Policy</a>
-                <span class="text-red-500 ml-1">*</span>
-              </label>
-            </div>
-            <div *ngIf="isFieldInvalid('terms')" class="text-sm text-red-600 -mt-2">
-              {{ getFieldError('terms') }}
-            </div>
-
-            <!-- Marketing Emails -->
-            <div class="flex items-start gap-3">
-              <div class="relative">
-                <input
-                  id="marketing"
-                  type="checkbox"
-                  formControlName="marketing"
-                  class="sr-only"
-                >
-                <div
-                  class="w-5 h-5 border-2 rounded transition-all duration-200 flex items-center justify-center cursor-pointer"
-                  [class]="signupForm.get('marketing')?.value
-                    ? 'bg-blue-600 border-blue-600'
-                    : 'bg-white border-gray-300 hover:border-blue-400'"
-                  (click)="toggleMarketing()"
-                >
-                  <svg *ngIf="signupForm.get('marketing')?.value" class="w-3 h-3 text-white" fill="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-                    <path fill-rule="evenodd" d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z" clip-rule="evenodd"/>
-                  </svg>
-                </div>
-              </div>
-              <label for="marketing" class="text-sm text-gray-700 cursor-pointer">
-                Send me product updates and marketing emails
-              </label>
-            </div>
-
-            <!-- Error Message -->
-            <div *ngIf="errorMessage()" class="p-4 bg-red-50 border border-red-200 rounded-xl">
-              <div class="flex items-center gap-3">
-                <svg class="w-5 h-5 text-red-500 flex-shrink-0" fill="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
-                  <path fill-rule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clip-rule="evenodd"/>
-                </svg>
-                <p class="text-sm text-red-800">{{ errorMessage() }}</p>
-              </div>
-            </div>
-
-            <!-- Submit Button -->
-            <app-button
-              type="submit"
-              variant="primary"
-              size="lg"
-              class="w-full"
-              [disabled]="signupForm.invalid || isLoading()"
-              [loading]="isLoading()"
-            >
-              <svg class="w-5 h-5" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M18 9v3m0 0v3m0-3h3m-3 0h-3m-2-5a4 4 0 11-8 0 4 4 0 018 0zM3 20a6 6 0 0112 0v1H3v-1z"/>
-              </svg>
-              Create Account
-            </app-button>
-          </form>
-
-          <!-- Login Link -->
-          <div class="mt-8 text-center">
-            <p class="text-gray-600">
-              Already have an account?
-              <a routerLink="/login" class="text-blue-600 hover:text-blue-800 font-semibold transition-colors">
-                Sign in here
-              </a>
-            </p>
-          </div>
-        </div>
-
-        <!-- Social Login -->
-        <div class="mt-8">
-          <div class="relative">
-            <div class="absolute inset-0 flex items-center">
-              <div class="w-full border-t border-gray-300"></div>
-            </div>
-            <div class="relative flex justify-center text-sm">
-              <span class="px-4 bg-gray-50 text-gray-500">Or continue with</span>
-            </div>
-          </div>
-
-          <div class="mt-6 grid grid-cols-2 gap-3">
-            <button
-              type="button"
-              class="flex justify-center items-center gap-2 w-full px-4 py-3 border border-gray-300 rounded-xl shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-              (click)="signupWithGoogle()"
-            >
-              <svg class="w-5 h-5" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                <path fill="#4285F4" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-                <path fill="#34A853" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-                <path fill="#FBBC05" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-                <path fill="#EA4335" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-              </svg>
-              Google
-            </button>
-
-            <button
-              type="button"
-              class="flex justify-center items-center gap-2 w-full px-4 py-3 border border-gray-300 rounded-xl shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 transition-colors"
-              (click)="signupWithGitHub()"
-            >
-              <svg class="w-5 h-5" fill="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
-              </svg>
-              GitHub
-            </button>
-          </div>
-        </div>
-      </div>
-
-      <!-- Password Generator Modal -->
-      <app-password-generator
-        *ngIf="showPasswordGen()"
-        (passwordGenerated)="onPasswordGenerated($event)"
-        (modalClosed)="hidePasswordGenerator()"
-      />
-    </div>
-  `
-})
-export class SignupComponent {
-  signupForm: FormGroup;
-  currentPassword = signal('');
-  errorMessage = signal('');
-  isLoading = signal(false);
-  showPasswordGen = signal(false);
-
-  constructor(
-    private fb: FormBuilder,
-    private authService: AuthService,
-    private router: Router
-  ) {
-    this.signupForm = this.fb.group({
-      fullName: ['', [Validators.required, Validators.minLength(2)]],
-      email: ['', [Validators.required, Validators.email]],
-      password: ['', [
-        Validators.required,
-        Validators.minLength(8),
-        Validators.pattern(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]/)
-      ]],
-      confirmPassword: ['', [Validators.required]],
-      terms: [false, [Validators.requiredTrue]],
-      marketing: [false]
-    }, {
-      validators: this.passwordMatchValidator
-    });
-  }
-
-  private passwordMatchValidator(form: any) {
-    const password = form.get('password');
-    const confirmPassword = form.get('confirmPassword');
-
-    if (password && confirmPassword && password.value !== confirmPassword.value) {
-      confirmPassword.setErrors({ passwordMismatch: true });
-      return { passwordMismatch: true };
-    }
-
-    return null;
-  }
-
-  isFieldInvalid(fieldName: string): boolean {
-    const field = this.signupForm.get(fieldName);
-    return !!(field?.invalid && (field?.dirty || field?.touched));
-  }
-
-  getFieldError(fieldName: string): string {
-    const field = this.signupForm.get(fieldName);
-    if (field?.errors && (field.dirty || field.touched)) {
-      if (field.errors['required']) return `${this.getFieldLabel(fieldName)} is required`;
-      if (field.errors['email']) return 'Please enter a valid email address';
-      if (field.errors['minlength']) return `${this.getFieldLabel(fieldName)} must be at least ${field.errors['minlength'].requiredLength} characters`;
-      if (field.errors['pattern'] && fieldName === 'password') return 'Password must contain uppercase, lowercase, number, and special character';
-      if (field.errors['passwordMismatch']) return 'Passwords do not match';
-      if (field.errors['requiredTrue']) return 'You must agree to the terms and conditions';
-    }
-    return '';
-  }
-
-  private getFieldLabel(fieldName: string): string {
-    const labels: { [key: string]: string } = {
-      fullName: 'Full name',
-      email: 'Email address',
-      password: 'Password',
-      confirmPassword: 'Confirm password'
-    };
-    return labels[fieldName] || fieldName;
-  }
-
-  onPasswordChange(event: Event): void {
-    const target = event.target as HTMLInputElement;
-    this.currentPassword.set(target.value);
-  }
-
-  toggleTerms(): void {
-    const current = this.signupForm.get('terms')?.value;
-    this.signupForm.patchValue({ terms: !current });
-  }
-
-  toggleMarketing(): void {
-    const current = this.signupForm.get('marketing')?.value;
-    this.signupForm.patchValue({ marketing: !current });
-  }
-
-  showPasswordGenerator(): void {
-    this.showPasswordGen.set(true);
-  }
-
-  hidePasswordGenerator(): void {
-    this.showPasswordGen.set(false);
-  }
-
-  onPasswordGenerated(password: string): void {
-    this.signupForm.patchValue({
-      password: password,
-      confirmPassword: password
-    });
-    this.currentPassword.set(password);
-    this.hidePasswordGenerator();
-  }
-
-  async onSubmit(): Promise<void> {
-    if (this.signupForm.invalid) {
-      this.signupForm.markAllAsTouched();
-      return;
-    }
-
-    this.isLoading.set(true);
-    this.errorMessage.set('');
-
-    try {
-      const formValue = this.signupForm.value;
-      await this.authService.signup({
-        name: formValue.fullName,
-        email: formValue.email,
-        password: formValue.password
-      });
-
-      this.router.navigate(['/dashboard']);
-    } catch (error: any) {
-      this.errorMessage.set(error.message || 'An error occurred during signup. Please try again.');
-    } finally {
-      this.isLoading.set(false);
-    }
-  }
-
-  async signupWithGoogle(): Promise<void> {
-    // Implement Google OAuth signup
-    console.log('Signup with Google');
-  }
-
-  async signupWithGitHub(): Promise<void> {
-    // Implement GitHub OAuth signup
-    console.log('Signup with GitHub');
-  }
-}
-
