+++ src/app/features/auth/components/login/login.component.ts
@@ -1,0 +1,102 @@
+import { Component, signal } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
+import { Router, RouterModule } from '@angular/router';
+import { AuthService } from '../../services/auth.service';
+import { FormInputComponent } from '../../../../shared/ui/form-input.component';
+import { ButtonComponent } from '../../../../shared/ui/button.component';
+
+@Component({
+  selector: 'app-login',
+  standalone: true,
+  imports: [
+    CommonModule,
+    ReactiveFormsModule,
+    RouterModule,
+    FormInputComponent,
+    ButtonComponent
+  ],
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.css']
+})
+export class LoginComponent {
+  loginForm: FormGroup;
+  errorMessage = signal('');
+  isLoading = signal(false);
+
+  constructor(
+    private fb: FormBuilder,
+    private authService: AuthService,
+    private router: Router
+  ) {
+    this.loginForm = this.fb.group({
+      email: ['', [Validators.required, Validators.email]],
+      password: ['', [Validators.required, Validators.minLength(6)]],
+      rememberMe: [false]
+    });
+  }
+
+  isFieldInvalid(fieldName: string): boolean {
+    const field = this.loginForm.get(fieldName);
+    return !!(field?.invalid && (field?.dirty || field?.touched));
+  }
+
+  getFieldError(fieldName: string): string {
+    const field = this.loginForm.get(fieldName);
+    if (field?.errors && (field.dirty || field.touched)) {
+      if (field.errors['required']) return `${this.getFieldLabel(fieldName)} is required`;
+      if (field.errors['email']) return 'Please enter a valid email address';
+      if (field.errors['minlength']) return `${this.getFieldLabel(fieldName)} must be at least ${field.errors['minlength'].requiredLength} characters`;
+    }
+    return '';
+  }
+
+  private getFieldLabel(fieldName: string): string {
+    const labels: { [key: string]: string } = {
+      email: 'Email address',
+      password: 'Password'
+    };
+    return labels[fieldName] || fieldName;
+  }
+
+  toggleRememberMe(): void {
+    const current = this.loginForm.get('rememberMe')?.value;
+    this.loginForm.patchValue({ rememberMe: !current });
+  }
+
+  async onSubmit(): Promise<void> {
+    if (this.loginForm.invalid) {
+      this.loginForm.markAllAsTouched();
+      return;
+    }
+
+    this.isLoading.set(true);
+    this.errorMessage.set('');
+
+    try {
+      const formValue = this.loginForm.value;
+      await this.authService.login({
+        email: formValue.email,
+        password: formValue.password,
+        rememberMe: formValue.rememberMe
+      });
+
+      this.router.navigate(['/dashboard']);
+    } catch (error: any) {
+      this.errorMessage.set(error.message || 'Invalid email or password. Please try again.');
+    } finally {
+      this.isLoading.set(false);
+    }
+  }
+
+  async loginWithGoogle(): Promise<void> {
+    // Implement Google OAuth login
+    console.log('Login with Google');
+  }
+
+  async loginWithGitHub(): Promise<void> {
+    // Implement GitHub OAuth login
+    console.log('Login with GitHub');
+  }
+}
+
