+++ src/app/features/auth/components/signup/signup.component.css
@@ -1,0 +1,4 @@
+/*
+  Component-specific styles can be placed here.
+  However, this component is currently styled using Tailwind CSS utility classes directly in the HTML template.
+*/
