+++ src/app/features/landing-page/components/landing.component.css
@@ -1,0 +1,11 @@
+:host {
+  display: block;
+}
+
+main {
+  scroll-margin-top: 80px;
+}
+
+section {
+  scroll-margin-top: 80px;
+}
