+++ src/app/features/landing-page/components/landing.component.html
@@ -1,0 +1,23 @@
+<app-header></app-header>
+
+<main class="pt-20">
+  <app-hero></app-hero>
+
+  <section id="features">
+    <app-features></app-features>
+  </section>
+
+  <section id="how-it-works">
+    <app-how-it-works></app-how-it-works>
+  </section>
+
+  <section id="testimonials">
+    <app-testimonials></app-testimonials>
+  </section>
+
+  <section id="pricing">
+    <app-pricing></app-pricing>
+  </section>
+</main>
+
+<app-footer></app-footer>
