--- src/app/header.component.css
@@ -1,103 +1,0 @@
-header {
-  background-color: #282c34;
-  padding: 10px;
-  color: white;
-  text-align: center;
-  border-bottom: 2px solid var(--pill-accent);
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-h1 {
-  margin: 0;
-  font-weight: 700;
-  line-height: 1.2;
-  color: var(--white);
-  text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
-  display: flex;
-  flex-direction: column;
-}
-h1 span {
-  color: maroon;
-  font-size: 1.5rem;
-}
-
-nav a {
-  color: var(--white);
-  text-decoration: none;
-  margin: 0 15px;
-  font-size: 1.2rem;
-  cursor: pointer;
-}
-nav a:hover {
-  text-decoration: underline;
-  color: teal;
-  transition: color 0.3s ease;
-}
-nav a.active {
-  font-weight: bold;
-  color: var(--pill-accent);
-}
-nav ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-}
-
-.search-bar {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.search-bar input {
-  padding: 10px;
-  font-size: 1rem;
-  border: 1px solid var(--gray-400);
-  border-radius: 5px;
-  width: 300px;
-  outline: none;
-}
-
-.search-bar button {
-  padding: 10px 20px;
-  font-size: 1rem;
-  background-color: green;
-  color: var(--white);
-  border: none;
-  border-radius: 5px;
-  margin-left: 10px;
-  cursor: pointer;
-}
-
-.user-profile {
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  display: flex;
-  gap: 3px;
-  padding: 10px;
-}
-
-.user-initials {
-  width: 50px;
-  height: 50px;
-  background-color: #092f7c;
-  color: var(--white);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  font-size: 1.5rem;
-}
-.username {
-  color: rgb(65, 3, 207);
-  font-size: 1.1rem;
-}
-
