--- src/app/header.component.html
@@ -1,29 +1,0 @@
-<header>
-    <!-- logo -->
-    <h1>
-        Unravel
-        <span>Docs</span>
-    </h1>
-
-    <!-- nav links -->
-    <nav>
-        <ul>
-            <li><a routerLink="/">Home</a></li>
-            <li><a routerLink="/about">About</a></li>
-            <li><a routerLink="/docs">Docs</a></li>
-            <li><a routerLink="/contact">Contact</a></li>
-        </ul>
-    </nav>
-    <!-- search bar -->
-    <div class="search-bar">
-        <input type="text" placeholder="Search..." />
-        <button type="submit">Search</button>
-    </div>
-    <!-- user profile -->
-    <div class="user-profile">
-        <div class="user-initials">
-            <span>JD</span>
-        </div>
-        <span class="username">John Doe</span>
-    </div>
-</header>
