--- src/app/header.component.ts
@@ -1,252 +1,0 @@
-import { Component, inject } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { Router } from '@angular/router';
-import { AuthService } from './services/auth.service';
-
-@Component({
-  selector: 'app-header',
-  standalone: true,
-  imports: [CommonModule],
-  template: `
-    <header class="fixed top-0 left-0 right-0 z-50 transition-all duration-300"
-            [class]="isScrolled ? 'bg-white/95 backdrop-blur-lg shadow-lg' : 'bg-white/90 backdrop-blur-md'">
-      <div class="max-w-7xl mx-auto px-6">
-        <div class="flex items-center justify-between h-20">
-          <!-- Logo -->
-          <div class="flex items-center cursor-pointer" (click)="navigateHome()">
-            <div class="h-10 w-10 bg-gradient-to-br from-blue-500 to-blue-700 rounded-lg flex items-center justify-center">
-              <span class="text-white font-bold text-lg">UD</span>
-            </div>
-            <span class="ml-3 text-xl font-bold text-gray-900">UnravelDocs</span>
-          </div>
-
-          <!-- Desktop Navigation -->
-          <nav class="hidden md:flex items-center space-x-8">
-            <button class="nav-link" *ngIf="!authService.isAuthenticated" (click)="scrollToSection('features')">Features</button>
-            <button class="nav-link" *ngIf="!authService.isAuthenticated" (click)="scrollToSection('how-it-works')">How It Works</button>
-            <button class="nav-link" *ngIf="!authService.isAuthenticated" (click)="scrollToSection('pricing')">Pricing</button>
-            <button class="nav-link" *ngIf="!authService.isAuthenticated" (click)="scrollToSection('testimonials')">Testimonials</button>
-
-            <!-- Authenticated Navigation -->
-            <a href="/dashboard" class="nav-link" *ngIf="authService.isAuthenticated">Dashboard</a>
-            <a href="/documents" class="nav-link" *ngIf="authService.isAuthenticated">My Documents</a>
-          </nav>
-
-          <!-- Auth Buttons - Not Authenticated -->
-          <div class="hidden md:flex items-center space-x-4" *ngIf="!authService.isAuthenticated">
-            <button class="px-5 py-2 text-gray-700 hover:text-gray-900 font-medium transition-colors"
-                    (click)="navigateToLogin()">
-              Log In
-            </button>
-            <button class="btn-primary group" (click)="navigateToSignup()">
-              Sign Up Free
-              <svg class="w-4 h-4 group-hover:translate-x-1 transition-transform" fill="none" stroke="currentColor"
-                   xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M13 7l5 5m0 0l-5 5m5-5H6"/>
-              </svg>
-            </button>
-          </div>
-
-          <!-- User Menu - Authenticated -->
-          <div class="hidden md:flex items-center space-x-4" *ngIf="authService.isAuthenticated">
-            <!-- User Plan Badge -->
-            <span class="px-3 py-1 text-xs font-semibold rounded-full"
-                  [class]="getPlanBadgeClasses(authService.currentUser()?.plan || 'free')">
-              {{ getPlanLabel(authService.currentUser()?.plan || 'free') }}
-            </span>
-
-            <!-- User Dropdown -->
-            <div class="relative" [class.open]="userMenuOpen">
-              <button class="flex items-center space-x-3 p-2 rounded-lg hover:bg-gray-100 transition-colors"
-                      (click)="toggleUserMenu()">
-                <div
-                  class="w-8 h-8 bg-gradient-to-br from-blue-500 to-blue-700 rounded-full flex items-center justify-center text-white font-semibold text-sm">
-                  {{ getInitials(authService.currentUser()?.name || 'User') }}
-                </div>
-                <span class="text-gray-700 font-medium">{{ authService.currentUser()?.name }}</span>
-                <svg class="w-4 h-4 text-gray-500 transition-transform" [class.rotate-180]="userMenuOpen" fill="none"
-                     stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                  <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"/>
-                </svg>
-              </button>
-
-              <!-- Dropdown Menu -->
-              <div class="user-dropdown" [class.show]="userMenuOpen">
-                <a href="/profile" class="dropdown-item">
-                  <svg class="w-4 h-4" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                          d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z"/>
-                  </svg>
-                  Profile
-                </a>
-                <a href="/settings" class="dropdown-item">
-                  <svg class="w-4 h-4" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                          d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z"/>
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                          d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"/>
-                  </svg>
-                  Settings
-                </a>
-                <a href="/billing" class="dropdown-item">
-                  <svg class="w-4 h-4" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                          d="M3 10h18M7 15h1m4 0h1m-7 4h12a3 3 0 003-3V8a3 3 0 00-3-3H6a3 3 0 00-3 3v8a3 3 0 003 3z"/>
-                  </svg>
-                  Billing
-                </a>
-                <div class="border-t border-gray-200 my-2"></div>
-                <button class="dropdown-item text-red-600 hover:bg-red-50 w-full text-left" (click)="logout()">
-                  <svg class="w-4 h-4" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                          d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"/>
-                  </svg>
-                  Sign Out
-                </button>
-              </div>
-            </div>
-          </div>
-
-          <!-- Mobile Menu Button -->
-          <button class="md:hidden flex flex-col justify-center items-center w-8 h-8 space-y-1"
-                  (click)="toggleMobileMenu()"
-                  [class.open]="mobileMenuOpen">
-            <span class="mobile-menu-line"></span>
-            <span class="mobile-menu-line"></span>
-            <span class="mobile-menu-line"></span>
-          </button>
-        </div>
-
-        <!-- Mobile Navigation -->
-        <nav class="md:hidden transition-all duration-300 overflow-hidden"
-             [class]="mobileMenuOpen ? 'max-h-96 opacity-100' : 'max-h-0 opacity-0'">
-          <div class="py-4 space-y-2 border-t border-gray-200">
-            <!-- Non-authenticated mobile menu -->
-            <div *ngIf="!authService.isAuthenticated">
-              <button class="mobile-nav-link w-full text-left" (click)="scrollToSection('features'); closeMobileMenu()">Features</button>
-              <button class="mobile-nav-link w-full text-left" (click)="scrollToSection('how-it-works'); closeMobileMenu()">How It Works</button>
-              <button class="mobile-nav-link w-full text-left" (click)="scrollToSection('pricing'); closeMobileMenu()">Pricing</button>
-              <button class="mobile-nav-link w-full text-left" (click)="scrollToSection('testimonials'); closeMobileMenu()">Testimonials</button>
-              <div class="pt-4 space-y-2 border-t border-gray-200 mt-4">
-                <button class="w-full text-left py-2 text-gray-700 font-medium" (click)="navigateToLogin()">Log In
-                </button>
-                <button class="w-full btn-primary justify-center" (click)="navigateToSignup()">Sign Up Free</button>
-              </div>
-            </div>
-
-            <!-- Authenticated mobile menu -->
-            <div *ngIf="authService.isAuthenticated">
-              <div class="flex items-center space-x-3 p-4 bg-gray-50 rounded-lg mb-4">
-                <div
-                  class="w-10 h-10 bg-gradient-to-br from-blue-500 to-blue-700 rounded-full flex items-center justify-center text-white font-semibold">
-                  {{ getInitials(authService.currentUser()?.name || 'User') }}
-                </div>
-                <div>
-                  <div class="font-semibold text-gray-900">{{ authService.currentUser()?.name }}</div>
-                  <div class="text-sm text-gray-600">{{ authService.currentUser()?.email }}</div>
-                </div>
-              </div>
-              <a href="/dashboard" class="mobile-nav-link" (click)="closeMobileMenu()">Dashboard</a>
-              <a href="/documents" class="mobile-nav-link" (click)="closeMobileMenu()">My Documents</a>
-              <a href="/profile" class="mobile-nav-link" (click)="closeMobileMenu()">Profile</a>
-              <a href="/settings" class="mobile-nav-link" (click)="closeMobileMenu()">Settings</a>
-              <button class="mobile-nav-link text-red-600 w-full text-left" (click)="logout()">Sign Out</button>
-            </div>
-          </div>
-        </nav>
-      </div>
-    </header>
-  `
-})
-export class HeaderComponent {
-  private router = inject(Router);
-  authService = inject(AuthService);
-
-  isScrolled = false;
-  mobileMenuOpen = false;
-  userMenuOpen = false;
-
-  constructor() {
-    if (typeof window !== 'undefined') {
-      window.addEventListener('scroll', () => {
-        this.isScrolled = window.scrollY > 50;
-      });
-
-      // Close user menu when clicking outside
-      document.addEventListener('click', (event) => {
-        if (!event.target || !(event.target as Element).closest('.relative')) {
-          this.userMenuOpen = false;
-        }
-      });
-    }
-  }
-
-  navigateHome(): void {
-    this.router.navigate(['/']);
-  }
-
-  navigateToLogin(): void {
-    this.closeMobileMenu();
-    this.router.navigate(['/login']);
-  }
-
-  navigateToSignup(): void {
-    this.closeMobileMenu();
-    this.router.navigate(['/signup']);
-  }
-
-  toggleMobileMenu(): void {
-    this.mobileMenuOpen = !this.mobileMenuOpen;
-  }
-
-  toggleUserMenu(): void {
-    this.userMenuOpen = !this.userMenuOpen;
-  }
-
-  closeMobileMenu(): void {
-    this.mobileMenuOpen = false;
-  }
-
-  scrollToSection(sectionId: string): void {
-    if (typeof document !== 'undefined') {
-      const element = document.getElementById(sectionId);
-      if (element) {
-        element.scrollIntoView({ behavior: 'smooth' });
-      }
-    }
-  }
-
-  logout(): void {
-    this.authService.logout().then(r => {
-      // Optionally handle post-logout actions here
-    });
-    this.userMenuOpen = false;
-    this.closeMobileMenu();
-    this.router.navigate(['/']).then(r => {
-      // Optionally handle post-navigation actions here
-    });
-  }
-
-  getPlanBadgeClasses(plan: string): string {
-    const classes = {
-      free: 'bg-gray-100 text-gray-800',
-      pro: 'bg-blue-100 text-blue-800',
-      enterprise: 'bg-purple-100 text-purple-800'
-    };
-    return classes[plan as keyof typeof classes] || classes.free;
-  }
-
-  getPlanLabel(plan: string): string {
-    const labels: { [key: string]: string } = {
-      free: 'Free',
-      pro: 'Pro',
-      enterprise: 'Enterprise'
-    };
-    return labels[plan] || 'Free';
-  }
-
-  getInitials(name: string): string {
-    if (!name) return 'U';
-    return name.split(' ').map(n => n.charAt(0)).join('').toUpperCase().slice(0, 2);
-  }
-}
