+++ src/app/shared/components/hero.component.ts
