+++ src/app/shared/components/how-it-works.component.ts
