+++ src/app/shared/components/navbar/footer/footer.component.css
@@ -1,0 +1,5 @@
+.social-link {
+  @apply w-10 h-10 bg-white/10 rounded-lg flex items-center justify-center
+  text-gray-400 hover:text-white hover:bg-blue-600 transition-all duration-300
+  hover:-translate-y-1;
+}
