+++ src/app/shared/components/navbar/footer/footer.component.html
@@ -1,0 +1,67 @@
+<footer class="bg-gray-900 text-white py-16">
+  <div class="max-w-7xl mx-auto px-6">
+    <div class="grid md:grid-cols-2 lg:grid-cols-5 gap-12 mb-12">
+      <!-- Brand Section -->
+      <div class="lg:col-span-2">
+        <img src="assets/logo.svg" alt="UnravelDocs" class="h-10 mb-6 brightness-0 invert">
+        <p class="text-gray-400 leading-relaxed max-w-sm mb-8">
+          Transform your documents with AI-powered text extraction.
+          Fast, accurate, and secure document digitization for professionals.
+        </p>
+
+        <!-- Social Links -->
+        <div class="flex gap-4">
+          <a href="#" class="social-link" aria-label="Twitter">
+            <svg viewBox="0 0 24 24" class="w-5 h-5 fill-current">
+              <path d="M23 3a10.9 10.9 0 01-3.14 1.53 4.48 4.48 0 00-7.86 3v1A10.66 10.66 0 013 4s-4 9 5 13a11.64 11.64 0 01-7 2c9 5 20 0 20-11.5a4.5 4.5 0 00-.08-.83A7.72 7.72 0 0023 3z"/>
+            </svg>
+          </a>
+          <a href="#" class="social-link" aria-label="LinkedIn">
+            <svg viewBox="0 0 24 24" class="w-5 h-5 fill-current">
+              <path d="M16 8a6 6 0 016 6v7h-4v-7a2 2 0 00-2-2 2 2 0 00-2 2v7h-4v-7a6 6 0 016-6zM2 9h4v12H2z"/>
+              <circle cx="4" cy="4" r="2"/>
+            </svg>
+          </a>
+          <a href="#" class="social-link" aria-label="GitHub">
+            <svg viewBox="0 0 24 24" class="w-5 h-5 fill-current">
+              <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
+            </svg>
+          </a>
+        </div>
+      </div>
+
+      <!-- Link Groups -->
+      @for (group of linkGroups; track group) {
+        <div class="space-y-4">
+          <h4 class="font-semibold text-white mb-4">{{ group.title }}</h4>
+          <ul class="space-y-3">
+            @for (link of group.links; track link) {
+              <li>
+                <a [href]="link.href" class="text-gray-400 hover:text-white transition-colors">
+                  {{ link.text }}
+                </a>
+              </li>
+            }
+          </ul>
+        </div>
+      }
+    </div>
+
+    <!-- Bottom Section -->
+    <div class="border-t border-gray-800 pt-8 flex flex-col md:flex-row justify-between items-center gap-4">
+      <p class="text-gray-400 text-sm">
+        © {{ currentYear }} UnravelDocs. All rights reserved.
+      </p>
+
+      <div class="flex items-center gap-6 text-sm">
+            <span class="flex items-center gap-2 text-gray-400">
+              <svg class="w-4 h-4 text-green-500" viewBox="0 0 20 20 fill-current">
+                <path fill-rule="evenodd" d="M2.166 4.999A11.954 11.954 0 0010 1.944 11.954 11.954 0 0017.834 5c.11.65.166 1.32.166 2.001 0 5.225-3.34 9.67-8 11.317C5.34 16.67 2 12.225 2 7c0-.682.057-1.35.166-2.001zm11.541 3.708a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clip-rule="evenodd"/>
+              </svg>
+              SSL Secured
+            </span>
+        <span class="text-gray-400">GDPR Compliant</span>
+      </div>
+    </div>
+  </div>
+</footer>
