+++ src/app/shared/components/pricing.component.ts
