+++ src/app/shared/ui/password-generator.component.ts
