+++ src/app/shared/ui/password-strength.component.ts
